split formula per Emacs version
