+++ .github/workflows/emacs-28.yml
@@ -5,11 +5,11 @@ on:
     branches:
       - master
     paths:
-      - 'Formula/emacs-plus.rb'
+      - 'Formula/emacs-plus@28.rb'
       - 'patches/**'
   pull_request:
     paths:
-      - 'Formula/emacs-plus.rb'
+      - 'Formula/emacs-plus@28.rb'
       - 'patches/**'
   schedule:
     - cron:  '0 */4 * * *'
@@ -21,9 +21,10 @@ jobs:
     strategy:
       matrix:
         build_opts:
-          - '--HEAD'
-          - '--HEAD --with-xwidgets'
-          - '--HEAD --with-no-titlebar'
+          - ''
+          - '--with-xwidgets'
+          - '--with-no-titlebar'
+          - '--with-x11'
 
     env:
       HOMEBREW_GITHUB_REF: ${{ github.head_ref || github.ref }}
@@ -33,13 +34,13 @@ jobs:
     steps:
     - uses: actions/checkout@v1
 
-    - name: Build emacs-plus ${{ matrix.build_opts }}
-      run: brew install Formula/emacs-plus.rb ${{ matrix.build_opts }}
+    - name: Build emacs-plus@28 ${{ matrix.build_opts }}
+      run: brew install Formula/emacs-plus@28.rb ${{ matrix.build_opts }}
 
     - name: Test installation
       if: contains(matrix.build_opts, '--HEAD') == false
-      run: brew test Formula/emacs-plus.rb
+      run: brew test Formula/emacs-plus@28.rb
 
     - name: Test installation (--HEAD)
       if: contains(matrix.build_opts, '--HEAD')
-      run: brew test Formula/emacs-plus.rb --HEAD
+      run: brew test Formula/emacs-plus@28.rb --HEAD
