+++ patches/emacs-26/fix-window-role.patch
