+++ patches/emacs-26/multicolor-fonts.patch
