+++ patches/emacs-27/no-titlebar.patch
