+++ patches/emacs-27/system-appearance.patch
