+++ patches/emacs-28/system-appearance.patch
